Collision detection
Added a display() and isHit() method to Boss. Changed enemyWidth and enemyHeight in Enemy to floats. Worked on bossCollision() in Projectile and added an update() method  for Projectile and Boss.

+++ Main/Enemy.pde
@@ -9,8 +9,8 @@ class Enemy
   private float xPos;
   private float yPos;
   private int typeOfGun;
-  private int enemyWidth;
-  private int enemyHeight;
+  private float enemyWidth;
+  private float enemyHeight;
   private int hitPoints;
   private PImage enemyPic;
   
@@ -52,12 +52,12 @@ class Enemy
     return yPos;
   }
   
-  public int getEnemyHeight()
+  public float getEnemyHeight()
   {
     return enemyHeight;
   }
   
-  public int getEnemyWidth()
+  public float getEnemyWidth()
   {
     return enemyWidth;
   }
